% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelationHeat.R
\name{plotCorrelationHeat}
\alias{plotCorrelationHeat}
\title{plotCorrelationHeat}
\usage{
plotCorrelationHeat(gs, markers = NULL, corMethod = "pearson",
  colors = colorRampPalette(rev(c("orangered4", "orangered", "gray90",
  "dodgerblue", "dodgerblue4")))(n = 100), outs = FALSE,
  displayNumbers = TRUE)
}
\arguments{
\item{gs}{The output of \code{sortGenes}.}

\item{markers}{Restrict correlation analysis to those genes. A character
vector.}

\item{corMethod}{Correlation method, will passed to \code{method} in the
function \code{cor}.}

\item{colors}{Color palette for drawing the heatmap}

\item{outs}{Should the \code{pheatmap} object and correlation matrix be 
returned? FALSE by default.}

\item{displayNumbers}{Should correlation values be displayed on the heatmap?
TRUE by default.}
}
\value{
If \code{outs} is TRUE, the pheatmap object and the correlation matrix will 
be returned.
}
\description{
\code{plotCorrelationHeat} uses the specificity scores returned by \code{sortGenes} to 
correlate cell clusters with each other and plot a heatmap.
}
\examples{
data(kidneyTabulaMuris)
gs = sortGenes(kidneyTabulaMuris$exp, kidneyTabulaMuris$cellType)
plotCorrelationHeat(gs)

#user only marker genes and spearman correlation
mm = getMarkers(gs, quant = 0.95)
plotCorrelationHeat(gs, markers = mm$markers, corMethod = "spearman")

#do not write correlation values, useful if there are many cell clusters
plotCorrelationHeat(gs, markers = mm$markers, corMethod = "spearman", displayNumbers = FALSE)
}
\author{
Mahmoud M Ibrahim <mmibrahim@pm.me>
}
