% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopBinaryHeat.R
\name{plotTopBinaryHeat}
\alias{plotTopBinaryHeat}
\title{plotTopBinaryHeat}
\usage{
plotTopBinaryHeat(sg, top_n = 10, colors = colorRampPalette(c("white",
  "black"))(n = 100), newOrder = 1:length(unique(sg$inputClass)),
  averageCells = 0, gaps = TRUE, outs = FALSE, plotheat = TRUE)
}
\arguments{
\item{sg}{A list, typically the output of \code{sortGenes}.}

\item{top_n}{The number of top genes to plot for each cluster.}

\item{colors}{Color palette used for the heatmap.}

\item{newOrder}{Reorder the clusters in the heatmap? See Examples.}

\item{averageCells}{Plot averages of cells instead of individual cells. You 
can use this when you have a large number of cells. See Details.}

\item{gaps}{Should the heatmap have gaps between cell types and gene clusters? 
TRUE by default.}

\item{outs}{Should the top genes names be returned? FALSE by default.}

\item{plotheat}{Should the heatmap be drawn? TRUE by default.}
}
\value{
If \code{outs} is TRUE, \code{plotMarkerHeat} returns a list containing 
the top \code{n} marker genes for each cluster.
}
\description{
Plot a heatmap of binarized values for the top genes or peaks in each cluster, 
defined by \code{sortGenes()}.
}
\details{
\code{plotTopMarkerHeat} is similar to \code{plotTopMarkerHeat} but works on 
the binarized matrix, plotting only 0s and 1s (for scRNA-Seq, this means ignoring 
expression values and just plotting whether the gene is expressed or not). 
If \code{averageCells} is > 1, the fraction of the averaged cells where the gene 
or peak was detected will be plotted (a number between 0 and 1).
}
\examples{
data(kidneyTabulaMuris)
gs = sortGenes(kidneyTabulaMuris$exp, kidneyTabulaMuris$cellType)
plotTopBinaryHeat(gs) # plots the top 10 genes for each cluster

#now plot the top 20 genes and average every 5 cells
plotTopBinaryHeat(gs, top_n= 20, averageCells=5)

}
\seealso{
plotTopMarkerHeat
}
\author{
Mahmoud M Ibrahim <mmibrahim@pm.me>
}
