% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPValues.R
\name{getPValues}
\alias{getPValues}
\title{getPValues}
\usage{
getPValues(gs, numPerm = 5, correctMethod = "BH", testGenes = NULL,
  subsetCells = NULL, cores = 1, seed = 111)
}
\arguments{
\item{gs}{The output of \code{sortGenes()}.}

\item{numPerm}{The number of permutations to do. The default value 5 is to
ensure quick running time for very large datasets but can be increased to
increase the power of the permutation test, see Details.}

\item{correctMethod}{The method used to correct p-values for multiple
hypothesis testing. Any valid input to "method" in \code{p.adjust} is
allowed. p-value correction is done on a gene-by-gene basis.}

\item{testGenes}{A character vector of gene names to restrict the output
p-values to.}

\item{subsetCells}{A numeric vector of cell indeces to restrict the
permutation to. Note that the selected cells should still contain at least
one cell from each of the cell clusters contained in \code{$specScore}.
Note this option is still experimental and might not give consistent
results.}

\item{cores}{An integer greater than zero (1 by default) that indicates how
many cores to use for parallelization using mclapply.}

\item{seed}{The seed for random permutations.}
}
\value{
\code{getPValues} returns a list with the following components:
  \item{permuteVal}{A sparse matrix with as many rows as genes and as many
  columns as \code{ncol($specScore) * numPerm}, containing the specificity
  score matrices for all permutations concatenated after each other
  column-wise.} \item{startIndeces}{A numeric vector of length \code{numPerm}
  that indicates the starting column for each performed permutation in
  \code{permuteVal}} \item{pval}{A matrix with as many rows as genes and
  columns as \code{ncol($specScore)}, containing the p-values for the null
  hypothesis that the gene is not highly specific to a cell cluster is true.}
  \item{adjpval}{A matrix with the same size as \code{pval}, containing the
  corrected p-values using the method specified in \code{correctMethod}}
}
\description{
getPValues performs a permutation test on the gene-cell type specificity
score to obtain a p-value value on each gene-cell type specificity score. The
permutation keeps everything the same except that cell type assignments are
permuted between the cells (but note that cell type proportions are also kept
the same). This function is a way to select all differentially expressed
genes between all cell classes globally in a dataset in one go or to define 
differentially expressed genes in a specific cell cluster.
}
\examples{
data(kidneyTabulaMuris)
gs = sortGenes(kidneyTabulaMuris$exp, kidneyTabulaMuris$cellType)
pp = getPValues(gs)
#obtain genes that are "differentially expressed" in at least one cluster
markers = names(which(apply(pp$adjpval, 1, function(x) any(x < 0.01))))

}
\seealso{
getMarkers
}
\author{
Mahmoud M Ibrahim <mmibrahim@pm.me>
}
