% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarkerScores.R
\name{plotMarkerScores}
\alias{plotMarkerScores}
\title{plotMarkerScores}
\usage{
plotMarkerScores(mm, gs, markers = mm$markers, colors = c("#99999920",
  "orangered4"))
}
\arguments{
\item{mm}{A list. The output of \code{getMarkers}}

\item{gs}{A list. The output of \code{sortGenes}}

\item{markers}{A character vector containing the genes to highlight in the
scatter plot.}

\item{colors}{A color palette to use for the scatter plot.}
}
\description{
plotMarkerScores plots scatter plots of the specificity score and the gene
specificity Shannon index to investigate \code{getMarkers} results.
}
\examples{
data(kidneyTabulaMuris)
sg = sortGenes(kidneyTabulaMuris$exp, kidneyTabulaMuris$cellType)
mm = getMarkers(sg,quant=0.99)
plotMarkerScores(mm, sg)
}
\author{
Mahmoud M Ibrahim <m3i@selfishscience.net>
}
